Eliminate DefaultDecoderNames
